Upgrade JSON Toolkit to make use of the new default resolver
