+++ src/cli/command_canonicalize.cc
@@ -1,6 +1,5 @@
 #include "commands.h"
 #include "defaults.h"
-#include "resolver.h"
 
 #include <jsonbinpack/canonicalizer/canonicalizer.h>
 #include <jsontoolkit/json.h>
@@ -14,7 +13,7 @@
 static auto canonicalize_from_json(sourcemeta::jsontoolkit::JSON &schema)
     -> int {
   sourcemeta::jsonbinpack::Canonicalizer canonicalizer{
-      sourcemeta::jsonbinpack::cli::Resolver{}};
+      sourcemeta::jsontoolkit::DefaultResolver{}};
   canonicalizer.apply(schema, sourcemeta::jsonbinpack::DEFAULT_METASCHEMA);
   sourcemeta::jsontoolkit::prettify(schema, std::cout);
   std::cout << std::endl;
