+++ test/canonicalizer/2020_12_boolean_test.cc
@@ -1,14 +1,10 @@
-#include "resolver.h"
-
 #include <jsonbinpack/canonicalizer/canonicalizer.h>
 #include <jsontoolkit/json.h>
 
 #include <gtest/gtest.h>
 
-// A shared instance for the tests
-static Resolver resolver{};
-
 TEST(CanonicalizerBoolean_2020_12, type_boolean) {
+  sourcemeta::jsontoolkit::DefaultResolver resolver;
   sourcemeta::jsonbinpack::Canonicalizer canonicalizer{resolver};
 
   sourcemeta::jsontoolkit::JSON schema{sourcemeta::jsontoolkit::parse(R"JSON({
@@ -28,6 +24,7 @@ TEST(CanonicalizerBoolean_2020_12, type_boolean) {
 }
 
 TEST(CanonicalizerBoolean_2020_12, drop_non_boolean_keywords_1) {
+  sourcemeta::jsontoolkit::DefaultResolver resolver;
   sourcemeta::jsonbinpack::Canonicalizer canonicalizer{resolver};
 
   sourcemeta::jsontoolkit::JSON schema{sourcemeta::jsontoolkit::parse(R"JSON({
@@ -49,6 +46,7 @@ TEST(CanonicalizerBoolean_2020_12, drop_non_boolean_keywords_1) {
 }
 
 TEST(CanonicalizerBoolean_2020_12, drop_non_boolean_keywords_2) {
+  sourcemeta::jsontoolkit::DefaultResolver resolver;
   sourcemeta::jsonbinpack::Canonicalizer canonicalizer{resolver};
 
   sourcemeta::jsontoolkit::JSON schema{sourcemeta::jsontoolkit::parse(R"JSON({
@@ -70,6 +68,7 @@ TEST(CanonicalizerBoolean_2020_12, drop_non_boolean_keywords_2) {
 }
 
 TEST(CanonicalizerBoolean_2020_12, drop_non_boolean_keywords_3) {
+  sourcemeta::jsontoolkit::DefaultResolver resolver;
   sourcemeta::jsonbinpack::Canonicalizer canonicalizer{resolver};
 
   sourcemeta::jsontoolkit::JSON schema{sourcemeta::jsontoolkit::parse(R"JSON({
