+++ test/canonicalizer/CMakeLists.txt
@@ -1,6 +1,5 @@
 add_executable(sourcemeta_jsonbinpack_canonicalizer_unit
   "${PROJECT_SOURCE_DIR}/vendor/googletest/googletest/src/gtest_main.cc"
-  resolver.h
   canonicalizer_test.cc
   2020_12_any_test.cc
   2020_12_array_test.cc
