+++ test/mapper/mapper_test.cc
@@ -1,15 +1,11 @@
-#include "resolver.h"
-
 #include <jsonbinpack/mapper/mapper.h>
 #include <jsontoolkit/json.h>
 
 #include <gtest/gtest.h>
 #include <stdexcept>
 
-// A shared instance for the tests
-static Resolver resolver{};
-
 TEST(Mapper, unsupported_dialect) {
+  sourcemeta::jsontoolkit::DefaultResolver resolver;
   sourcemeta::jsonbinpack::Mapper mapper{resolver};
   sourcemeta::jsontoolkit::JSON schema{sourcemeta::jsontoolkit::parse(R"JSON({
     "$schema": "https://json-schema.org/draft/2019-09/schema",
@@ -20,6 +16,7 @@ TEST(Mapper, unsupported_dialect) {
 }
 
 TEST(Mapper, unknown_dialect) {
+  sourcemeta::jsontoolkit::DefaultResolver resolver;
   sourcemeta::jsonbinpack::Mapper mapper{resolver};
   sourcemeta::jsontoolkit::JSON schema{sourcemeta::jsontoolkit::parse(R"JSON({
     "type": "boolean"
